+++ src/services/app_service.py
@@ -22,6 +22,20 @@ class AppService:
         self._user = None
 
     def login(self, username, password):
+        """Login a user
+
+        Checks if the username and password match and logs the user in
+
+        Args:
+            username (str):
+                The username to log in
+            password (str):
+                The password to log in
+
+        Returns:
+            True if the user was logged in, False if the username and password do not match
+        """
+
         success = ur.check_password(username, password)
 
         if not success:
@@ -32,6 +46,7 @@ class AppService:
         return True
 
     def logout(self):
+        """Logout the current user"""
         self._user = None
 
     def create_user(self, username, password):
@@ -55,7 +70,39 @@ class AppService:
             return False
 
         ur.add_user(username, password)
+        user_id = ur.get_id(username)[0]
+
+        ur.init_new_user_stat(user_id)
+
         return True
+
+    def save_stats(self):
+        """Save the user stats to the database"""
+
+        if not self._user:
+            return
+
+        user_id = ur.get_id(self._user)[0]
+        stats = ur.get_user_stats(user_id)
+        score = self._game.get_total_score()
+
+        if score > stats[2]:
+            ur.update_top_score(user_id, score)
+
+        ur.update_games_played(user_id, stats[3] + 1)
+
+        best_on_rows = self.get_best_hand_rows()
+        best_on_columns = self.get_best_hand_columns()
+        hands_made = stats[4].split(';')
+
+        for i in range(10):
+            last = int(hands_made[i])
+            row_count = best_on_rows.count(i)
+            column_count = best_on_columns.count(i)
+            hands_made[i] = str(last + row_count + column_count)
+
+        hands_made = ';'.join(hands_made)
+        ur.update_hands_made(user_id, hands_made)
 
     # Game related methods
 
@@ -94,6 +141,21 @@ class AppService:
 
     def get_hand_score(self, hand):
         return self._game.get_score(hand)
+
+    def get_user(self):
+        return self._user
+
+    def get_id(self, user):
+        return ur.get_id(user)
+
+    def get_user_stats(self, user_id):
+        return ur.get_user_stats(user_id)
+
+    def get_top_scores(self):
+        return ur.get_top_scores()
+
+    def get_user_by_id(self, user_id):
+        return ur.get_user_by_id(user_id)
 
 
 app_service = AppService()