+++ dokumentaatio/changelog.md
@@ -35,3 +35,7 @@
 - Commented the code with Docstring style
 - Added manual
 - Updated Architecture doc
+
+## Week 7
+- added GameoverView where stats are shown for the player after gameover
+- added user stats saving to database